+++ app/models/burden.rb
@@ -1,5 +1,6 @@
 class Burden < ActiveHash::Base
-  self.date = [
+  self.data = [
+    {id: 0, name: '---'},
     {id: 1, name: '着払い(購入者負担)'},
     {id: 2, name: '送料込み(出品者負担)'}
   ]
