+++ app/models/origin.rb
@@ -1,5 +1,6 @@
 class Origin < ActiveHash::Base
-  self.date = [
+  self.data = [
+    {id: 0, name: '---'},
     {id: 1, name: '北海道'},
     {id: 2, name: '青森県'},
     {id: 3, name: '岩手県'},
